My - Added CRUD in product #1

+++ Repositories/Entities/GoldPrice.cs
@@ -11,8 +11,8 @@ namespace Repositories.Entities
         public Gold Gold { get; set; }
         public DateTime DateTime { get; set; }
         [Column(TypeName = "decimal(18,2)")]
-        public float AskPrice { get; set; }
+        public decimal AskPrice { get; set; }
         [Column(TypeName = "decimal(18,2)")]
-        public float BidPrice { get; set; }
+        public decimal BidPrice { get; set; }
     }
 }
